# openapi-ai.yaml
openapi: 3.0.0
info:
  title: UniMate AI 模块
  version: 1.0.0
  description: 学习规划AI助手接口

servers:
  - url: http://localhost:8080
    description: 本地开发环境

tags:
  - name: AI助手
    description: AI对话、学习计划生成、目标拆解等功能

paths:
  /health:
    get:
      summary: 健康检查
      tags:
        - 系统
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: AI 服务运行中

  /api/ai/chat:
    post:
      summary: AI对话
      description: 与AI进行对话，支持上下文记忆功能
      tags:
        - AI助手
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - message
              properties:
                user_id:
                  type: string
                  description: 用户ID
                  example: AN051
                message:
                  type: string
                  description: 用户消息内容
                  example: 我想准备考研数学，不知道从哪里开始
                conversation_id:
                  type: string
                  description: 对话ID（续接对话时需要，首次对话不需要）
                  example: user_AN051_conv_1762594131
            examples:
              首次对话:
                value:
                  user_id: AN051
                  message: 我想准备考研数学
              续接对话:
                value:
                  user_id: AN051
                  message: 我数学基础比较差怎么办？
                  conversation_id: user_AN051_conv_1762594131
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      reply:
                        type: string
                        description: AI的回复内容
                        example: 你好！准备考研数学需要系统规划...
                      conversation_id:
                        type: string
                        description: 对话ID，用于续接对话
                        example: user_AN051_conv_1762594131
                      token_usage:
                        type: object
                        description: Token使用情况和费用统计
                        properties:
                          input:
                            type: integer
                            description: 输入Token数
                            example: 132
                          output:
                            type: integer
                            description: 输出Token数
                            example: 51
                          total:
                            type: integer
                            description: 总Token数
                            example: 183
                          cost:
                            type: number
                            format: double
                            description: 本次调用费用（人民币）
                            example: 0.00006405
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: 参数错误

  /api/ai/generate-plan:
    post:
      summary: 生成学习计划
      description: 根据学习目标、截止日期等信息，AI生成详细的分阶段学习计划
      tags:
        - AI助手
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - goal
              properties:
                user_id:
                  type: string
                  description: 用户ID
                  example: AN051
                goal:
                  type: string
                  description: 学习目标
                  example: 准备考研数学，目标130分以上
                deadline:
                  type: string
                  description: 截止日期
                  example: "2025-12-31"
                current_level:
                  type: string
                  description: 当前水平描述
                  example: 基础薄弱，高数只有60分
                daily_hours:
                  type: integer
                  description: 每天可学习小时数
                  example: 3
            example:
              user_id: AN051
              goal: 准备考研数学，目标130分以上
              deadline: "2025-12-31"
              current_level: 基础薄弱
              daily_hours: 3
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      plan_id:
                        type: string
                        description: 计划ID
                        example: plan_1762594200
                      title:
                        type: string
                        description: 计划标题
                        example: 考研数学130分学习计划
                      phases:
                        type: array
                        description: 学习阶段
                        items:
                          type: object
                          properties:
                            phase_name:
                              type: string
                              description: 阶段名称
                              example: 第一阶段：基础夯实
                            duration:
                              type: string
                              description: 阶段时长
                              example: 1-2个月
                            tasks:
                              type: array
                              description: 具体任务列表
                              items:
                                type: string
                              example:
                                - 完成高等数学教材第1-5章
                                - 做课后习题巩固基础
                                - 总结重要知识点和公式
                      suggestions:
                        type: array
                        description: 学习建议
                        items:
                          type: string
                        example:
                          - 建议每天固定时间学习，养成习惯
                          - 多做练习题，注重理解而非死记硬背
                      token_usage:
                        type: object
                        properties:
                          input:
                            type: integer
                            example: 200
                          output:
                            type: integer
                            example: 300
                          total:
                            type: integer
                            example: 500
                          cost:
                            type: number
                            example: 0.000175

  /api/ai/breakdown-goal:
    post:
      summary: 目标拆解
      description: 将大目标拆解为3-6个具体的、可执行的子任务
      tags:
        - AI助手
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - goal
              properties:
                user_id:
                  type: string
                  description: 用户ID
                  example: AN051
                goal:
                  type: string
                  description: 主目标
                  example: 通过英语六级考试，目标500分以上
                deadline:
                  type: string
                  description: 截止日期
                  example: "2025-06-30"
            example:
              user_id: AN051
              goal: 通过英语六级考试，目标500分以上
              deadline: "2025-06-30"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      main_goal:
                        type: string
                        description: 主目标
                        example: 通过英语六级考试，目标500分以上
                      sub_goals:
                        type: array
                        description: 子任务列表
                        items:
                          type: object
                          properties:
                            task:
                              type: string
                              description: 具体任务描述
                              example: 每天背诵50个六级核心词汇
                            estimated_days:
                              type: integer
                              description: 预估完成天数
                              example: 90
                            priority:
                              type: string
                              enum: [高, 中, 低]
                              description: 优先级
                              example: 高
                      token_usage:
                        type: object
                        properties:
                          total:
                            type: integer
                            example: 350
                          cost:
                            type: number
                            example: 0.0001225

  /api/ai/evaluate-plan:
    post:
      summary: 评估计划
      description: 评估学习计划的可行性，给出评分、优缺点和优化建议
      tags:
        - AI助手
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - goal
                - plan_content
              properties:
                user_id:
                  type: string
                  description: 用户ID
                  example: AN051
                goal:
                  type: string
                  description: 学习目标
                  example: 学会Python数据分析
                plan_content:
                  type: string
                  description: 计划内容
                  example: 第一周学语法，第二周学Pandas，第三周做项目，第四周找工作
                daily_hours:
                  type: integer
                  description: 每天可学习小时数
                  example: 2
                deadline:
                  type: string
                  description: 截止日期
                  example: "2025-03-31"
            example:
              user_id: AN051
              goal: 学会Python数据分析
              plan_content: 第一周学语法，第二周学Pandas，第三周做项目，第四周找工作
              daily_hours: 2
              deadline: "2025-03-31"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      score:
                        type: integer
                        description: 可行性评分（0-100）
                        example: 75
                      strengths:
                        type: array
                        description: 计划优点
                        items:
                          type: string
                        example:
                          - 时间安排合理，循序渐进
                          - 有明确的阶段划分
                      issues:
                        type: array
                        description: 存在的问题
                        items:
                          type: string
                        example:
                          - 缺少练习巩固环节
                          - 项目难度可能过高
                      suggestions:
                        type: array
                        description: 优化建议
                        items:
                          type: string
                        example:
                          - 建议在第二周增加每日练习时间
                          - 项目可以从简单的数据清洗开始
                      token_usage:
                        type: object
                        properties:
                          total:
                            type: integer
                            example: 400
                          cost:
                            type: number
                            example: 0.00014

  /api/ai/conversations:
    get:
      summary: 获取对话列表
      description: 获取指定用户的所有对话记录
      tags:
        - AI助手
      parameters:
        - name: user_id
          in: query
          required: true
          description: 用户ID
          schema:
            type: string
            example: AN051
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      conversations:
                        type: array
                        items:
                          type: object
                          properties:
                            conversation_id:
                              type: string
                              example: user_AN051_conv_1762594131
                            message_count:
                              type: integer
                              description: 消息数量
                              example: 5
                            last_message:
                              type: string
                              description: 最后一条消息
                              example: 我想准备考研数学
                            created_at:
                              type: string
                              description: 创建时间
                              example: "2025-01-08T10:00:00Z"

  /api/ai/conversations/{conversation_id}:
    delete:
      summary: 删除对话
      description: 删除指定的对话记录
      tags:
        - AI助手
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: 对话ID
          schema:
            type: string
            example: user_AN051_conv_1762594131
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 对话已删除
        '404':
          description: 对话不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: 对话不存在